fetch comments, add logging capabitilies, check bug with anorm and None types

+++ app/model/dtos/DiscussionThread.scala
@@ -1,8 +1,8 @@
 package model.dtos
 
-case class DiscussionThread(id:BigInt,
-                            from:Byte,
-                            to:Byte,
+case class DiscussionThread(id:Long,
+                            from:Int,
+                            to:Int,
                             body:String,
                             numberOfComments:Int,
                             numberOfAnnotations:Int) //a comment may have zero or more annotations attached
